requirements.txt
numpy~=1.26.2
opencv-python~=4.5.4.60
cvzone~=1.6.1
face-recognition
firebase-admin
cvzone
datetime~=5.4
flask~=3.0.0
Pillow~=10.1.0
dlib~=19.24.2
pip~=23.3.2
wheel~=0.42.0
cryptography~=41.0.7
Jinja2~=3.1.2
setuptools~=58.1.0
packaging~=23.2
zipp~=3.17.0
build~=1.0.3